# cmd/docs/swagger.yaml
definitions:
...
  github_com_sankangkin_di-rest-api_internal_models.Customer:
    properties:
      address:
        minLength: 3
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        minLength: 3
        type: string
      phone:
        minLength: 3
        type: string
      updatedAt:
        type: string
    required:
    - address
    - name
    - phone
    type: object
  github_com_sankangkin_di-rest-api_internal_models.Inventory:
    properties:
      createdAt:
        type: string
      createdTime:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      inQty:
        type: integer
      outQty:
        type: integer
      product:
        $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Product'
      productId:
        type: string
      remark:
        type: string
      updatedAt:
        type: string
      updatedTime:
        type: string
    type: object
...
  github_com_sankangkin_di-rest-api_internal_models.Product:
    properties:
      brand:
        type: string
      buyPrice:
        minimum: 1
        type: integer
      categoryId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      isActive:
        type: boolean
      productName:
        minLength: 3
        type: string
      sellPricelvl1:
        minimum: 1
        type: integer
      sellPricelvl2:
        minimum: 1
        type: integer
      uom:
        minLength: 3
        type: string
      uomId:
        minimum: 3
        type: integer
      updatedAt:
        type: string
    required:
    - buyPrice
    - productName
    - sellPricelvl1
    - sellPricelvl2
    - uom
    - uomId
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  httputil.HttpError400:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  httputil.HttpError401:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized
        type: string
    type: object
  httputil.HttpError500:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: internal server error
        type: string
    type: object
...
  internal_domain_customer.CreateCustomerRequestDTO:
    properties:
      address:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  internal_domain_customer.UpdateCustomerRequstDTO:
    properties:
      address:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  internal_domain_inventory.IncreaseInventoryDTO:
    properties:
      inQty:
        type: integer
      outQty:
        type: integer
      productId:
        type: string
      remark:
        type: string
    type: object
...
host: localhost:5555
info:
  contact:
    email: sankaungkin@gmail.com
    name: San KaungKin
  description: This is an auto-generated API docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: REST-API with(golang fiber, google wire dependency injection)
  version: "1.0"
paths:
...
  /api/customers:
    get:
      consumes:
      - application/json
      description: Fetch all customers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Customer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch all customers
      tags:
      - Customers
    post:
      consumes:
      - application/json
      description: Create new customer based on parameters
      parameters:
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_domain_customer.CreateCustomerRequestDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Create new customer based on parameters
      tags:
      - Customers
  /api/customers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete individual customer
      parameters:
      - description: customer Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Delete individual customer
      tags:
      - Customers
    get:
      consumes:
      - application/json
      description: Fetch individual customer by Id
      parameters:
      - description: customer Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch individual customer by Id
      tags:
      - Customers
    put:
      consumes:
      - application/json
      description: Update individual customer
      parameters:
      - description: customer Id
        in: path
        name: id
        required: true
        type: string
      - description: Product Data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_domain_customer.UpdateCustomerRequstDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Update individual customer
      tags:
      - Customers
  /api/inventories:
    get:
      consumes:
      - application/json
      description: Fetch all inventory records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Inventory'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Fetch all inventory records
      tags:
      - Inventories
  /api/inventories/decrease:
    post:
      consumes:
      - application/json
      description: Create decrease inventory record based on parameters
      parameters:
      - description: Inventory Data
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/internal_domain_inventory.IncreaseInventoryDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Inventory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Create decrease inventory record based on parameters
      tags:
      - Inventories
  /api/inventories/increase:
    post:
      consumes:
      - application/json
      description: Create increase inventory record based on parameters
      parameters:
      - description: Inventory Data
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/internal_domain_inventory.IncreaseInventoryDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Inventory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - Bearer: []
      summary: Create increase inventory record based on parameters
      tags:
      - Inventories
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"